Remove the development version tests
These have been failing for reasons I don't entirely understand.
The workflow obtains the development versions of qiskit-machine-learning
and qiskit-aer, but installs the latest stable version of
qiskit-terra for some reason.

I'm disabling until I can get to the bottom of this, and will open
a draft pull request that re-enables these tests.